Split out Questions component and TiviaHome screen, added progress indicator
